Backend: Split quick_links for layering/avoiding circular module refs
